requirements.txt
# NovaTech AI — Requirements
# Compatible with Streamlit Cloud (Python 3.11+)
# All packages are available on PyPI

# ── Core ─────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
python-dotenv>=1.0.0

# ── Data ─────────────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0

# ── ML ───────────────────────────────────────────────────────────────────────
scikit-learn>=1.4.0
joblib>=1.3.0

# ── Visualisation ────────────────────────────────────────────────────────────
plotly>=5.20.0
matplotlib>=3.8.0
Pillow>=10.0.0

# ── NLP ──────────────────────────────────────────────────────────────────────
nltk>=3.8.0

# ── Gemini AI ────────────────────────────────────────────────────────────────
google-generativeai>=0.7.0

# ── Translation (optional — graceful fallback if absent) ─────────────────────
deep-translator>=1.11.0

# ── Animation export ─────────────────────────────────────────────────────────
# Pillow handles GIF writing via matplotlib's PillowWriter
# No separate moviepy needed unless you want MP4 export

# ── Dev only (not needed on Streamlit Cloud) ──────────────────────────────────
# jupyter>=1.0.0
# ipykernel>=6.0.0
